refactor(eth): move PubSubConfig struct to pubsub.go to improve logical grouping
