+++ eth/pubsub.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/OffchainLabs/prysm/v6/beacon-chain/p2p"
+	"github.com/OffchainLabs/prysm/v6/beacon-chain/p2p/encoder"
 	ethtypes "github.com/OffchainLabs/prysm/v6/proto/prysm/v1alpha1"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	pubsubpb "github.com/libp2p/go-libp2p-pubsub/pb"
@@ -21,6 +22,36 @@ import (
 )
 
 const eventTypeHandleMessage = "HANDLE_MESSAGE"
+
+// PubSubConfig holds the configuration for pubsub event processing
+type PubSubConfig struct {
+	Topics         []string
+	ForkVersion    ForkVersion
+	Encoder        encoder.NetworkEncoding
+	SecondsPerSlot time.Duration
+	GenesisTime    time.Time
+	DataStream     host.DataStream
+}
+
+func (p PubSubConfig) Validate() error {
+	if p.Encoder == nil {
+		return fmt.Errorf("nil encoder")
+	}
+
+	if p.SecondsPerSlot == 0 {
+		return fmt.Errorf("seconds per slot must not be 0")
+	}
+
+	if p.GenesisTime.IsZero() {
+		return fmt.Errorf("genesis time must not be zero time")
+	}
+
+	if p.DataStream == nil {
+		return fmt.Errorf("datastream implementation required")
+	}
+
+	return nil
+}
 
 type PubSub struct {
 	host *host.Host
